requirements.txt
kaggle
python-dotenv
streamlit
langchain
openai
streamlit
streamlit-chat
watchdog
pytest
chromadb
tiktoken